update: replace fill_customer_cart with create_cart_items and move fixture to cart fixtures file
